+++ src/handlers/blog.handler.ts
@@ -1,9 +1,12 @@
 import { oakHandler } from "../types.ts";
+import { getBlogMetaList, getBlogPost } from "../utilities/parser.utility.ts";
 
 export const getBlogList: oakHandler<"/"> = async (ctx) => {
-  ctx.response.body = "index";
+  const metaList = await getBlogMetaList();
+  ctx.response.body = metaList;
 };
 
-export const getBlogById: oakHandler<"/:id"> = async (ctx) => {
-  ctx.response.body = ctx.params.id;
+export const getBlogByName: oakHandler<"/:name"> = async (ctx) => {
+  const blogData = await getBlogPost(ctx.params.name);
+  ctx.response.body = blogData;
 };