+++ src/routes/blog.router.ts
@@ -1,8 +1,8 @@
 import { oak } from "../deps.ts";
-import { getBlogById, getBlogList } from "../handlers/blog.handler.ts";
+import { getBlogByName, getBlogList } from "../handlers/blog.handler.ts";
 
 const blogRouter = new oak.Router({ prefix: "/blog" });
 
-blogRouter.get("/", getBlogList).get("/:id", getBlogById);
+blogRouter.get("/", getBlogList).get("/:name", getBlogByName);
 
 export default blogRouter;