+++ CHANGELOG.md
@@ -7,6 +7,13 @@ Addition of markdown posts will not be versioned or require a changelog update.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2022-5-12
+
+### Added
+
+- fixed handler type to correctly handle generic path params
+- wrote custom meta string parser that is very fast, great for the blog index page
+
 ## [0.2.5] - 2022-5-11
 
 ### Added